Tiny changes
git-svn-id: svn://svn.dolezel.info/fatrat/trunk/fatrat@94 78086a28-8538-0410-bde1-fa2ac3882955

+++ engines/TorrentProgressWidget.cpp
@@ -28,7 +28,7 @@ QImage TorrentProgressWidget::generate(const std::vector<bool>& data, int width,
 		int from = i*fact+sstart;
 		int to = (i+1)*fact+sstart;
 		
-		if(to >= data.size())
+		if(to >= (int) data.size())
 			to = data.size()-1;
 		
 		double color = 0;
